Refactored AdditionalResources into its own component. Updated link for CareerScope (#5450)
* Refactored AdditionalResources into its own component. Updated link for CareerScope.

* Renamed component.

* Added missing class from merge.
